// src/tasks/enums/tasks-messages.enum.ts
export enum TaskSuccessMessages {
  CREATED = "تسک ایجاد شد",
  FOUND = "تسک پیدا شد",
  LISTED = "لیست تسک ها دریافت شد",
  UPDATED = "تسک با موفقیت بروز رسانی شد",
  DELETED = "تسک با موفقیت حذف شد",
}

export enum TaskErrorMessages {
  DUPLICATE = "تسک تکراری است",
  NOT_FOUND = "تسک پیدا نشد",
}